+++ hardhat-config/providers.config.js
@@ -2,19 +2,13 @@ module.exports = {
   providers: {
     infura: {
       mainnet: '{{PROVIDER_URL_INFURA_MAINNET}}',
-      ropsten: '{{PROVIDER_URL_INFURA_ROPSTEN}}',
-      rinkeby: '{{PROVIDER_URL_INFURA_RINKEBY}}',
       goerli: '{{PROVIDER_URL_INFURA_GOERLI}}',
-      kovan: '{{PROVIDER_URL_INFURA_KOVAN}}',
       polygon: '{{PROVIDER_URL_INFURA_POLYGON}}',
       polygonMumbai: '{{PROVIDER_URL_INFURA_POLYGONMUMBAI}}',
     },
     alchemy: {
       mainnet: '{{PROVIDER_URL_ALCHEMY_MAINNET}}',
-      ropsten: '{{PROVIDER_URL_ALCHEMY_ROPSTEN}}',
-      rinkeby: '{{PROVIDER_URL_ALCHEMY_RINKEBY}}',
       goerli: '{{PROVIDER_URL_ALCHEMY_GOERLI}}',
-      kovan: '{{PROVIDER_URL_ALCHEMY_KOVAN}}',
       polygon: '{{PROVIDER_URL_ALCHEMY_POLYGON}}',
       polygonMumbai: '{{PROVIDER_URL_ALCHEMY_POLYGONMUMBAI}}',
     },