+++ migrations/token/ERC20/polygon_bridging/ERC20EscrowPredicate_deploy.js
@@ -1,0 +1,8 @@
+const Contract_deploy = require('../../../../src/templates/Contract/deploy');
+const {getNamedAccount} = require('../../../../src/helpers/templates');
+
+module.exports = Contract_deploy('ERC20EscrowPredicate', {
+  contract: 'ERC20EscrowPredicate',
+  args: [{name: 'RootChainManager', value: getNamedAccount('Polygon_RootChainManager_Proxy')}],
+});
+module.exports.tags = ['ERC20', 'PolygonPredicates'];